+++ ktor/client/src/commonMain/kotlin/dev/inmo/micro_utils/ktor/client/OnUploadCallback.kt
@@ -1,0 +1,3 @@
+package dev.inmo.micro_utils.ktor.client
+
+typealias OnUploadCallback = suspend (uploaded: Long, count: Long) -> Unit