+++ ktor/common/src/commonMain/kotlin/dev/inmo/micro_utils/ktor/common/InputProvider.kt
@@ -1,0 +1,5 @@
+package dev.inmo.micro_utils.ktor.common
+
+import io.ktor.utils.io.core.Input
+
+typealias LambdaInputProvider = () -> Input