+++ ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/ktor/server/QueryParameters.kt
@@ -1,0 +1,25 @@
+package dev.inmo.micro_utils.ktor.server
+
+import io.ktor.http.HttpStatusCode
+import io.ktor.server.application.ApplicationCall
+import io.ktor.server.response.respond
+
+suspend fun ApplicationCall.getParameterOrSendError(
+    field: String
+) = parameters[field].also {
+    if (it == null) {
+        respond(HttpStatusCode.BadRequest, "Request must contains $field")
+    }
+}
+
+fun ApplicationCall.getQueryParameter(
+    field: String
+) = request.queryParameters[field]
+
+suspend fun ApplicationCall.getQueryParameterOrSendError(
+    field: String
+) = getQueryParameter(field).also {
+    if (it == null) {
+        respond(HttpStatusCode.BadRequest, "Request query parameters must contains $field")
+    }
+}