+++ ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/ktor/server/UniloadMultipart.kt
@@ -2,22 +2,34 @@ package dev.inmo.micro_utils.ktor.server
 
 import dev.inmo.micro_utils.common.*
 import dev.inmo.micro_utils.coroutines.safely
-import dev.inmo.micro_utils.ktor.common.*
-import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.ApplicationCall
-import io.ktor.server.application.call
-import io.ktor.server.request.receive
 import io.ktor.server.request.receiveMultipart
-import io.ktor.server.response.respond
-import io.ktor.server.response.respondBytes
-import io.ktor.server.routing.Route
-import io.ktor.server.websocket.WebSocketServerSession
-import io.ktor.util.pipeline.PipelineContext
 import io.ktor.utils.io.core.*
-import kotlinx.coroutines.flow.Flow
-import kotlinx.serialization.DeserializationStrategy
-import kotlinx.serialization.SerializationStrategy
+import kotlinx.coroutines.currentCoroutineContext
+import kotlinx.coroutines.isActive
+
+/**
+ * Server-side part which receives [dev.inmo.micro_utils.ktor.client.uniUpload] request
+ */
+suspend inline fun ApplicationCall.handleUniUpload(
+    onFormItem: (PartData.FormItem) -> Unit = {},
+    onBinaryContent: (PartData.BinaryItem) -> Unit = {},
+    onBinaryChannelItem: (PartData.BinaryChannelItem) -> Unit = {},
+    onFileItem: (PartData.FileItem) -> Unit = {}
+) {
+    val multipartData = receiveMultipart()
+
+    while (currentCoroutineContext().isActive) {
+        val partData = multipartData.readPart() ?: break
+        when (partData) {
+            is PartData.FormItem -> onFormItem(partData)
+            is PartData.FileItem -> onFileItem(partData)
+            is PartData.BinaryItem -> onBinaryContent(partData)
+            is PartData.BinaryChannelItem -> onBinaryChannelItem(partData)
+        }
+    }
+}
 
 suspend fun ApplicationCall.uniloadMultipart(
     onFormItem: (PartData.FormItem) -> Unit = {},
@@ -25,21 +37,16 @@ suspend fun ApplicationCall.uniloadMultipart(
     onBinaryChannelItem: (PartData.BinaryChannelItem) -> Unit = {},
     onBinaryContent: (PartData.BinaryItem) -> Unit = {}
 ) = safely {
-    val multipartData = receiveMultipart()
-
     var resultInput: Input? = null
 
-    multipartData.forEachPart {
-        when (it) {
-            is PartData.FormItem -> onFormItem(it)
-            is PartData.FileItem -> {
-                when (it.name) {
-                    "bytes" -> resultInput = it.provider()
-                    else -> onCustomFileItem(it)
-                }
-            }
-            is PartData.BinaryItem -> onBinaryContent(it)
-            is PartData.BinaryChannelItem -> onBinaryChannelItem(it)
+    handleUniUpload(
+        onFormItem,
+        onBinaryContent,
+        onBinaryChannelItem
+    ) {
+        when (it.name) {
+            "bytes" -> resultInput = it.provider()
+            else -> onCustomFileItem(it)
         }
     }
 
@@ -88,24 +95,4 @@ suspend fun ApplicationCall.uniloadMultipartFile(
     }
 
     resultInput ?: error("Bytes has not been received")
-}
-
-suspend fun ApplicationCall.getParameterOrSendError(
-    field: String
-) = parameters[field].also {
-    if (it == null) {
-        respond(HttpStatusCode.BadRequest, "Request must contains $field")
-    }
-}
-
-fun ApplicationCall.getQueryParameter(
-    field: String
-) = request.queryParameters[field]
-
-suspend fun ApplicationCall.getQueryParameterOrSendError(
-    field: String
-) = getQueryParameter(field).also {
-    if (it == null) {
-        respond(HttpStatusCode.BadRequest, "Request query parameters must contains $field")
-    }
 }